requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
psycopg2-binary==2.9.7
configparser==6.0.0
webdriver-manager==4.0.1
brotli==1.1.0